auth_service/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
PyJWT==2.9.0
